+++ src/scippnexus/field.py
