+++ src/scippnexus/file.py
@@ -1,24 +1,33 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
 # @author Simon Heybrock
-import warnings
 from contextlib import AbstractContextManager
+from typing import Mapping
 
 import h5py
-from scipp import VisibleDeprecationWarning
 
-from .nexus_classes import NXroot
+from .base import Group, base_definitions
+
+_default_definitions = object()
 
 
-class File(AbstractContextManager, NXroot):
-    def __init__(self, *args, definition=None, **kwargs):
-        warnings.warn(
-            "This API is deprecated and will be removed and replaced in release 23.06. "
-            "Switch to 'import scippnexus.v2 as snx' to prepare for this.",
-            VisibleDeprecationWarning,
-        )
+class File(AbstractContextManager, Group):
+    def __init__(self, *args, definitions: Mapping = _default_definitions, **kwargs):
+        """Context manager for NeXus files, similar to h5py.File.
+
+        Arguments other than documented are as in :py:class:`h5py.File`.
+
+        Parameters
+        ----------
+        definitions:
+            Mapping of NX_class names to application-specific definitions.
+            The default is to use the base definitions as defined in the
+            NeXus standard.
+        """
+        if definitions is _default_definitions:
+            definitions = base_definitions()
         self._file = h5py.File(*args, **kwargs)
-        NXroot.__init__(self, self._file, definition=definition)
+        super().__init__(self._file, definitions=definitions)
 
     def __enter__(self):
         self._file.__enter__()
