+++ src/scippnexus/nexus_classes.py
@@ -1,18 +1,12 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
 # @author Simon Heybrock
+from .base import NXobject, NXroot  # noqa F401
 from .nxcylindrical_geometry import NXcylindrical_geometry  # noqa F401
-from .nxdata import NXdata  # noqa F401
-from .nxdetector import NXdetector  # noqa F401
-from .nxdisk_chopper import NXdisk_chopper  # noqa F401
+from .nxdata import NXdata, NXdetector, NXlog, NXmonitor  # noqa F401
 from .nxevent_data import NXevent_data  # noqa F401
-from .nxfermi_chopper import NXfermi_chopper  # noqa F401
-from .nxlog import NXlog  # noqa F401
-from .nxmonitor import NXmonitor  # noqa F401
-from .nxobject import NXobject, NXroot  # noqa F401
 from .nxoff_geometry import NXoff_geometry  # noqa F401
 from .nxsample import NXsample  # noqa F401
-from .nxsource import NXsource  # noqa F401
 from .nxtransformations import NXtransformations  # noqa F401
 
 
@@ -72,8 +66,16 @@ class NXdetector_module(NXobject):
     """NXdetector_module"""
 
 
+class NXdisk_chopper(NXobject):
+    """NXdisk_chopper"""
+
+
 class NXenvironment(NXobject):
     """NXenvironment"""
+
+
+class NXfermi_chopper(NXobject):
+    """NXfermi_chopper"""
 
 
 class NXfilter(NXobject):
@@ -162,6 +164,10 @@ class NXsensor(NXobject):
 
 class NXshape(NXobject):
     """NXshape"""
+
+
+class NXsource(NXobject):
+    """NXsource"""
 
 
 class NXslit(NXobject):
