+++ src/scippnexus/v1/__init__.py
@@ -11,15 +11,10 @@ except importlib.metadata.PackageNotFoundError:
     __version__ = "0.0.0"
 
 from .. import typing
-from .base import (
-    Group,
-    NexusStructureError,
-    NXobject,
-    base_definitions,
-    create_class,
-    create_field,
-)
-from .field import Attrs, Field
+from .definition import ApplicationDefinition, make_definition
 from .file import File
 from .nexus_classes import *
-from .nxdata import group_events_by_detector_number
+from .nxdata import NXdataStrategy
+from .nxdetector import NXdetectorStrategy
+from .nxlog import NXlogStrategy
+from .nxobject import Attrs, Field, NexusStructureError, NXobject, NXobjectStrategy
