+++ src/scippnexus/v1/definition.py
