+++ src/scippnexus/v1/definitions/nxcansas.py
