+++ src/scippnexus/v1/nxdetector.py
