+++ src/scippnexus/v1/nxdisk_chopper.py
