+++ src/scippnexus/v1/nxfermi_chopper.py
