+++ src/scippnexus/v1/nxlog.py
