+++ src/scippnexus/v1/nxmonitor.py
