+++ src/scippnexus/v1/nxsource.py
