+++ tests/nxtransformations_test.py
@@ -5,7 +5,7 @@ import scipp as sc
 from scipp.testing import assert_identical
 
 import scippnexus.v2 as snx
-from scippnexus.v2.nxtransformations import NXtransformations
+from scippnexus.nxtransformations import NXtransformations
 
 
 def make_group(group: h5py.Group) -> snx.Group:
